// emp-mgm/src/app/model/certificate.model.ts
export class Certificate{
    _id: string;
    coursename: string;
    courseid: string;
    coursecode: string;
    studentid: string;
    invid: string;
    name: string;
    email: string;
    telephone: string;
    date: Date;
    adminshow: Boolean;
    approved: Boolean;
    rejected: Boolean;
    certificateid:string;
}